fix building on Windows
just don't attempt to coordinate progress output for now
