+++ cmd/bass/progress.go
@@ -3,12 +3,9 @@ package main
 import (
 	"context"
 	"fmt"
-	"net"
 	"os"
 	"os/signal"
-	"syscall"
 
-	"github.com/adrg/xdg"
 	"github.com/morikuni/aec"
 	"github.com/opencontainers/go-digest"
 	"github.com/vito/bass/pkg/bass"
@@ -70,32 +67,4 @@ func withProgress(ctx context.Context, name string, f func(context.Context, *pro
 	}
 
 	return nil
-}
-
-func electRecorder() (ui.Reader, *progrock.Recorder, error) {
-	socketPath, err := xdg.StateFile(fmt.Sprintf("bass/recorder.%d.sock", syscall.Getpgrp()))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var r ui.Reader
-	var w progrock.Writer
-	l, err := net.Listen("unix", socketPath)
-	if err != nil {
-		r = nil // don't display any progress; send it to the leader
-		w, err = progrock.DialRPC("unix", socketPath)
-	} else {
-		r, w, err = progrock.ServeRPC(l)
-	}
-
-	return r, progrock.NewRecorder(w), err
-}
-
-func cleanupRecorder() error {
-	socketPath, err := xdg.StateFile(fmt.Sprintf("bass/recorder.%d.sock", syscall.Getpgrp()))
-	if err != nil {
-		return err
-	}
-
-	return os.RemoveAll(socketPath)
 }
